fix: Fix rounding issue on the buy order's fill
This issue arises when the buy order is being filled, in that case the
limiting amount multiplied by the order's price can be slightly less
than the actual amount left in the order because of the multiplication
following the division to calculate the limiting amount. This leaves some dust in the
order. To avoid this issue, when the buy order is filled just sent
everything left. This does not happen in the sell order because the
limiting amount is exactly the amount in the order